no longer a fork
while initially planned as a fork and rewrite of addui, this is no longer the case. I feel as though the codebase of addui is not very intuitive. The coding standard is also not to my liking, which- isnt to say that chinese coding is bad nessecarily, but that i prefer modularity similar to oop languages